==> code/0-prepare-data.ipynb <==
import numpy as np
import pandas as pd
from utils.dataset import *
from utils.text import clean_text

___
# Load cytology and histopathology reports
### Load metadata

# Define path to CHIFIR dataset
path = "../../../Data/CHIFIR/"

# Load a csv file with report IDs and labels
df = pd.read_csv(path + "/chifir_metadata.csv")
print(df.shape)
df.head()

### Load report texts

# Read report texts
df['order_results'] = df.apply(read_reports, path=path+"reports/", axis=1)

### Convert target variable to categorical

df.y_report = labels2cat(df.y_report)

### Summary statistics

print("Total number of reports:", df.shape[0])
print("Total number of unique patients:", df.patient_id.nunique())
print("Number of reports per patient varies from %d to %d with a median value of %d." % 
      (df.groupby('patient_id').size().min(), 
       df.groupby('patient_id').size().max(), 
       df.groupby('patient_id').size().median())
     )
print("The average document length is %d characters." % df.order_results.apply(len).mean())

print_stats(df, 'y_report')

### Examine data splits

# How many reports in dev and test sets?
df.dataset.value_counts(dropna=False)

# How many reports in each CV fold?
df.val_fold.value_counts().sort_index()

# Proportion of positive reports in each CV fold?
df.groupby('val_fold').y_report.value_counts(normalize=True).round(2)

### Parse files with gold standard annotations

# Map character positions before and after text pre-processing
df['pos_mapping'] = df.order_results.apply(clean_text, return_mapping=True)

read_annotations(df, path=path+"annotations/")

### Save datasets

df[df.dataset=='development'].to_csv("../datasets/reports_dev.csv", index=False)
df[df.dataset=='test'].to_csv("../datasets/reports_test.csv", index=False)
