==> code/2-gold-standard-eda.ipynb <==
import numpy as np
import pandas as pd
from utils.prod import get_feature_names
from utils import text

___
# EDA on the gold standard annotations
### Load data

# Define path to CHIFIR dataset
path = "../../../Data/CHIFIR/"

# Load metadata for the full / training / test set
df = pd.read_csv(path + "/chifir_metadata.csv")
print(df.shape)
df.head()

# Load gold standard annotations
concepts = text.load_annotations('concepts')
print(concepts.shape)
concepts.head()

relations = text.load_annotations('relations')
print(relations.shape)
relations.head()

composite_concepts = text.load_annotations('composite')
print(composite_concepts.shape)
composite_concepts.head()

# Filter to caluclate stats on dev and test sets
concepts = concepts.merge(df.histopathology_id)
relations = relations.merge(df.histopathology_id)
composite_concepts = composite_concepts.merge(df.histopathology_id)

concepts.shape, relations.shape, composite_concepts.shape

# Number of reports with no annotations
print("%d reports with no annotations of any kind." % 
      (~df.histopathology_id.isin(concepts.histopathology_id) & 
       ~df.histopathology_id.isin(relations.histopathology_id)).sum())

print("%d reports with no concepts." % (~df.histopathology_id.isin(concepts.histopathology_id)).sum())

print("%d reports with no relations." % (~df.histopathology_id.isin(relations.histopathology_id)).sum())

### Summary statistics for concepts

# Median number of concepts in the annotated reports
concepts.groupby('histopathology_id').size().median()

# Total occurences for each category
concepts.concept.value_counts().sort_index()

# Number of reports with at least one occurrence
concepts.groupby('concept', observed=False).histopathology_id.nunique()

# Number of unique phrases
concepts.groupby('concept', observed=False).phrase.nunique()

# Lexical diversity
(concepts.groupby('concept', observed=False).phrase.nunique() / concepts.concept.value_counts().sort_index()).round(2)

# Percentage of occurrences in positive reports
concepts = concepts.merge(df[['histopathology_id', 'y_report']])
concepts.groupby('concept', observed=False).y_report.value_counts(normalize=True).sort_index().round(2)

### Summary statistics for relations

# Median number of relations in the annotated reports
relations.groupby('histopathology_id').size().median()

# Total occurences for each category
relations.relation.value_counts().sort_index()

# Number of reports with at least one occurrence
relations.groupby('relation', observed=False).histopathology_id.nunique()

# Percentage of occurrences in positive reports
relations = relations.merge(df[['histopathology_id', 'y_report']])
relations.groupby('relation', observed=False).y_report.value_counts(normalize=True).sort_index().round(2)

# Total occurences for each category
composite_concepts.concept.value_counts().sort_index()

# Number of reports with at least one occurrence
composite_concepts.groupby('concept', observed=False).histopathology_id.nunique()

# Percentage of occurrences in positive reports
composite_concepts = composite_concepts.merge(df[['histopathology_id', 'y_report']])
composite_concepts.groupby('concept', observed=False).y_report.value_counts(normalize=True).sort_index().round(2)
