==> code/4-automated-extraction.ipynb <==
___
# Automated extraction of IFI-related information
___

import numpy as np
import pandas as pd
from utils import text, prod, eval_ner
from utils.dev import get_cv_strategy

# import warnings
# warnings.filterwarnings("ignore")

___
# Detect concepts and relations
### Run concept recognition and relationship detection on the development set with CV
**Load and prepare data**

# Load the development set
df = pd.read_csv("../datasets/reports_dev.csv")

# Clean data
df['clean_text'] = df.order_results.apply(text.clean_text)

**Run NER**

cv = get_cv_strategy()

for train_idx, val_idx in cv.split(df.clean_text, df.y_report, df.patient_id):
    
    # Learn vocabulary and termsets
    vocab = text.learn_vocab(df.loc[train_idx].histopathology_id, expand=True)
    termset = text.learn_termset(df.loc[train_idx].histopathology_id)
    
    # Load NLP pipeline
    nlp = text.build_nlp_pipeline(termset)
    
    # Match phrases in text to detect concepts
    df.loc[val_idx, 'doc'] = text.detect_concepts(df.loc[val_idx, 'clean_text'], nlp, vocab)

**Get detected concepts (either to use directly to evaluate NER or to get concept counts)**

# Transform predictions to a table of concepts
detected_concepts = text.get_concepts(df)

**Evaluate NER**

# Load gold standard concepts
concepts = text.load_annotations('composite')

# Get feature names
feature_names = prod.get_feature_names(['concepts','composite'])

# Calculate precision and recall
eval_ner.evaluate_ner(df[['histopathology_id', 'val_fold']], 
                      concepts, detected_concepts, feature_names)

### Run relationship detection using gold standard annotations on the development set with CV
**Load and prepare data**

# Load the development set
df = pd.read_csv("../datasets/reports_dev.csv")

# Clean data
df['clean_text'] = df.order_results.apply(text.clean_text)

**Run NER**

cv = get_cv_strategy()

for train_idx, val_idx in cv.split(df.clean_text, df.y_report, df.patient_id):
    
    # Learn termsets
    termset = text.learn_termset(df.loc[train_idx].histopathology_id)
    
    # Load NLP pipeline
    nlp = text.build_nlp_pipeline(termset)
    
    # Label known concepts
    df.loc[val_idx, 'doc'] = text.label_concepts(df.loc[val_idx, ['histopathology_id', 'clean_text']], nlp)


**Get detected concepts (either to use directly to evaluate NER or to get concept counts)**

# Transform predictions to a table of concepts
detected_concepts = text.get_concepts(df)

**Evaluate NER**

# Load gold standard concepts
concepts = text.load_annotations('composite')

# Get feature names
feature_names = prod.get_feature_names(['concepts','composite'])

# Calculate precision and recall
eval_ner.evaluate_ner(df[['histopathology_id', 'val_fold']], 
                      concepts, detected_concepts, feature_names)

### Run concept recognition and relationship detection on the test set

# Load the development set
df_test = pd.read_csv("../datasets/reports_test.csv")

# Clean data
df_test['clean_text'] = df_test.order_results.apply(text.clean_text)

# Learn vocabulary and termsets
vocab = text.learn_vocab(df.histopathology_id, expand=True)
termset = text.learn_termset(df.histopathology_id)

# Load NLP pipeline
nlp = text.build_nlp_pipeline(termset)

# Match phrases in text to detect concepts
df_test['doc'] = text.detect_concepts(df_test.clean_text, nlp, vocab)

# Transform predictions to a table of concepts
detected_concepts = text.get_concepts(df_test)

# Load gold standard concepts
concepts = text.load_annotations('composite')

# Get feature names
feature_names = prod.get_feature_names(['concepts','composite'])

# Calculate precision and recall
eval_ner.evaluate_ner(df_test[['histopathology_id', 'val_fold']], 
                      concepts, detected_concepts, feature_names)

### Run relationship detection using gold standard annotations on the test set

# Load the development set
df_test = pd.read_csv("../datasets/reports_test.csv")

# Clean data
df_test['clean_text'] = df_test.order_results.apply(text.clean_text)

# Learn termsets
termset = text.learn_termset(df.histopathology_id)

# Load NLP pipeline
nlp = text.build_nlp_pipeline(termset)

# Label known concepts
df_test['doc'] = text.label_concepts(df_test[['histopathology_id', 'clean_text']], nlp)

# Transform predictions to a table of concepts
detected_concepts = text.get_concepts(df_test)

# Load gold standard concepts
concepts = text.load_annotations('composite')

# Get feature names
feature_names = prod.get_feature_names(['concepts','composite'])

# Calculate precision and recall
eval_ner.evaluate_ner(df_test[['histopathology_id', 'val_fold']], 
                      concepts, detected_concepts, feature_names)
